refactor(components): refactor drawer
