+++ packages/components/drawer/src/drawer.vue
@@ -60,18 +60,13 @@
 <script lang="ts">
 import { defineComponent, computed, ref } from 'vue'
 import { ElOverlay } from '@element-plus/components/overlay'
-import {
-  useDialog,
-  dialogProps,
-  dialogEmits,
-} from '@element-plus/components/dialog'
+import { useDialog, dialogEmits } from '@element-plus/components/dialog'
 import ElIcon from '@element-plus/components/icon'
 import { TrapFocus } from '@element-plus/directives'
 import { Close } from '@element-plus/icons-vue'
+import { drawerProps } from './drawer'
 
-import type { PropType, SetupContext } from 'vue'
-
-type DrawerDirection = 'ltr' | 'rtl' | 'ttb' | 'btt'
+import type { SetupContext } from 'vue'
 
 export default defineComponent({
   name: 'ElDrawer',
@@ -83,29 +78,7 @@ export default defineComponent({
   directives: {
     TrapFocus,
   },
-  props: {
-    ...dialogProps,
-    direction: {
-      type: String as PropType<DrawerDirection>,
-      default: 'rtl',
-      validator: (val: DrawerDirection) => {
-        return ['ltr', 'rtl', 'ttb', 'btt'].indexOf(val) !== -1
-      },
-    },
-    size: {
-      type: [String, Number],
-      default: '30%',
-    },
-    withHeader: {
-      type: Boolean,
-      default: true,
-    },
-    modalFade: {
-      type: Boolean,
-      default: true,
-    },
-  },
-
+  props: drawerProps,
   emits: dialogEmits,
 
   setup(props, ctx) {
